chore: restructure time filter folders
