--- packages/instant-apps-components/src/testUtils/utils.ts
@@ -1,3 +1,0 @@
-export function isDate(obj: any): boolean {
-  return obj instanceof Array;
-}
